@@ -1,7 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+import 'package:jost_pay_wallet/Provider/DashboardProvider.dart';
+import 'package:jost_pay_wallet/Provider/theme_provider.dart';
+// import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:jost_pay_wallet/Values/MyColor.dart';
 import 'package:jost_pay_wallet/Values/utils.dart';
+import 'package:jost_pay_wallet/common/button.dart';
+import 'package:provider/provider.dart';
+import 'package:webview_flutter/webview_flutter.dart';
 
 class TermsScreen extends StatefulWidget {
   const TermsScreen({super.key});
@@ -11,36 +16,40 @@ class TermsScreen extends StatefulWidget {
 }
 
 class _TermsScreenState extends State<TermsScreen> {
-  late InAppWebViewController _webViewController;
   bool isLoading = true;
+  WebViewController? controller;
+  @override
+  void initState() {
+    super.initState();
+    controller = WebViewController()
+      ..setJavaScriptMode(JavaScriptMode.unrestricted)
+      ..setNavigationDelegate(
+        NavigationDelegate(
+          onProgress: (int progress) {
+            // Update loading bar.
+          },
+          onPageStarted: (String url) {},
+          onPageFinished: (String url) {
+            setState(() {
+              isLoading = false;
+            });
+          },
+          onHttpError: (HttpResponseError error) {},
+          onWebResourceError: (WebResourceError error) {},
+        ),
+      )
+      ..loadRequest(Uri.parse(Utils.termsUrl));
+  }
 
   @override
   Widget build(BuildContext context) {
+    final themeProvider = Provider.of<ThemeProvider>(context, listen: true);
+    final dashProvider = Provider.of<DashboardProvider>(context, listen: false);
     return Scaffold(
         body: SafeArea(
       child: Stack(
         children: [
-          InAppWebView(
-            initialUrlRequest: URLRequest(url: Uri.parse(Utils.termsUrl)),
-            initialOptions: InAppWebViewGroupOptions(
-              crossPlatform: InAppWebViewOptions(
-                useShouldOverrideUrlLoading: false,
-                javaScriptEnabled: true,
-                javaScriptCanOpenWindowsAutomatically: true,
-              ),
-            ),
-            onWebViewCreated: (controller) {
-              _webViewController = controller;
-            },
-            onLoadStop: (controller, url) async {
-              bool check = await controller.isLoading();
-              if (!check) {
-                setState(() {
-                  isLoading = false;
-                });
-              }
-            },
-          ),
+          WebViewWidget(controller: controller!),
           Visibility(
             visible: isLoading,
             child: const Center(
@@ -48,6 +57,41 @@ class _TermsScreenState extends State<TermsScreen> {
               color: MyColor.greenColor,
             )),
           ),
+          Positioned(
+              child: Container(
+            height: 68,
+            width: double.infinity,
+            color: themeProvider.isDarkMode()
+                ? MyColor.dark01Color
+                : MyColor.mainWhiteColor,
+            child: Row(
+              children: [
+                Transform.translate(
+                  offset: const Offset(10, -8),
+                  child: BackBtn(
+                    onTap: () {
+                      dashProvider.changeBottomIndex(4);
+                    },
+                  ),
+                ),
+                const Spacer(flex: 2),
+                Text(
+                  "Terms & Conditions",
+                  style: TextStyle(
+                    color: themeProvider.isDarkMode()
+                        ? MyColor.mainWhiteColor
+                        : MyColor.dark01Color,
+                    fontSize: 20,
+                    fontWeight: FontWeight.w500,
+                  ),
+                ),
+                const Spacer(
+                  flex: 3,
+                ),
+              ],
+            ),
+          ))
+       
         ],
       ),
     ));
