@@ -29,7 +29,40 @@ class _Pay4meSuccessScreenState extends State<Pay4meSuccessScreen> {
               const SizedBox(
                 height: 200,
               ),
-              Image.asset('assets/images/successB.png'),
+              Container(
+                  width: double.infinity,
+                  padding: EdgeInsets.all(10),
+                  decoration: BoxDecoration(
+                      color: Theme.of(context).colorScheme.tertiary,
+                      borderRadius: BorderRadius.circular(100),
+                      gradient: const LinearGradient(
+                        colors: [
+                          MyColor.greenColor,
+                          Color(0xff00c26d),
+                        ],
+                        begin: Alignment.topLeft,
+                        end: Alignment.bottomRight,
+                      )),
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Spacer(),
+                      Text(
+                        "Pay4me success!",
+                        style: MyStyle.tx16White.copyWith(
+                            fontWeight: FontWeight.bold, fontSize: 20),
+                      ),
+                      Spacer(),
+                      const CircleAvatar(
+                        radius: 20,
+                        backgroundColor: Colors.white,
+                        child: Icon(
+                          Icons.check,
+                          color: MyColor.greenColor,
+                        ),
+                      ),
+                    ],
+                  )),
               const SizedBox(
                 height: 40,
               ),
@@ -37,7 +70,8 @@ class _Pay4meSuccessScreenState extends State<Pay4meSuccessScreen> {
                 "Your Pay4Me order has been successfully placed. Our team is reviewing your request, and your invoice will be paid within 24 hours. You will receive a status update once the payment is completed",
                 textAlign: TextAlign.center,
                 style: MyStyle.tx14Grey.copyWith(
-                    color: Theme.of(context).colorScheme.tertiary, height: 2.5),
+                  color: Theme.of(context).colorScheme.tertiary,
+                ),
               ),
               const SizedBox(
                 height: 32,
@@ -45,12 +79,13 @@ class _Pay4meSuccessScreenState extends State<Pay4meSuccessScreen> {
               TextButton(
                 onPressed: () => Get.offAll(BottomNav()),
                 style: TextButton.styleFrom(
+                  padding: EdgeInsets.symmetric(horizontal: 30),
                   side: BorderSide(color: MyColor.greenColor),
                   shape: RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(10)),
                 ),
                 child: Text(
-                  'Go Homepage',
+                  'Got it',
                   style:
                       MyStyle.tx16Green.copyWith(fontWeight: FontWeight.w600),
                 ),
